facade for services, state machine more generic for K8s
