+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/kubernetes/facade/redis/config/KubernetesRedisConfig.groovy
@@ -1,4 +1,4 @@
-package com.swisscom.cloud.sb.broker.services.kubernetes.redis.config
+package com.swisscom.cloud.sb.broker.services.kubernetes.facade.redis.config
 
 import com.swisscom.cloud.sb.broker.cfextensions.endpoint.EndpointConfig
 import com.swisscom.cloud.sb.broker.config.Config
