+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/services/kubernetes/templates/KubernetesTemplateManagerSpec.groovy
@@ -1,6 +1,6 @@
 package com.swisscom.cloud.sb.broker.services.kubernetes.templates
 
-import com.swisscom.cloud.sb.broker.services.kubernetes.redis.config.KubernetesRedisConfig
+import com.swisscom.cloud.sb.broker.services.kubernetes.facade.redis.config.KubernetesRedisConfig
 import spock.lang.Specification
 
 class KubernetesTemplateManagerSpec extends Specification {
